+++ sdk/communication/communication-call-automation/test/callConnection.spec.ts
@@ -16,6 +16,12 @@ import {
   TransferCallResult,
   RemoveParticipantResult,
   MuteParticipantsResult,
+  CancelAddParticipantResult,
+  AddParticipantEventResult,
+  TransferCallToParticipantEventResult,
+  RemoveParticipantEventResult,
+  CancelAddParticipantEventResult,
+  AddParticipantCancelled,
 } from "../src";
 import Sinon, { SinonStubbedInstance } from "sinon";
 import { CALL_TARGET_ID, CALL_TARGET_ID_2 } from "./utils/connectionUtils";
@@ -156,7 +162,11 @@ describe("CallConnection Unit Tests", () => {
 
   it("AddParticipant", async () => {
     // mocks
-    const addParticipantResultMock: AddParticipantResult = {};
+    const addParticipantResultMock: AddParticipantResult = {
+      waitForEventProcessor: async () => {
+        return {} as AddParticipantEventResult;
+      },
+    };
     callConnection.addParticipant.returns(
       new Promise((resolve) => {
         resolve(addParticipantResultMock);
@@ -178,7 +188,11 @@ describe("CallConnection Unit Tests", () => {
 
   it("TransferCallToParticipant", async () => {
     // mocks
-    const transferCallResultMock: TransferCallResult = {};
+    const transferCallResultMock: TransferCallResult = {
+      waitForEventProcessor: async () => {
+        return {} as TransferCallToParticipantEventResult;
+      },
+    };
     callConnection.transferCallToParticipant.returns(
       new Promise((resolve) => {
         resolve(transferCallResultMock);
@@ -202,7 +216,11 @@ describe("CallConnection Unit Tests", () => {
 
   it("TransferCallToParticipantWithTransferee", async () => {
     // mocks
-    const transferCallResultMock: TransferCallResult = {};
+    const transferCallResultMock: TransferCallResult = {
+      waitForEventProcessor: async () => {
+        return {} as TransferCallToParticipantEventResult;
+      },
+    };
     callConnection.transferCallToParticipant.returns(
       new Promise((resolve) => {
         resolve(transferCallResultMock);
@@ -230,7 +248,11 @@ describe("CallConnection Unit Tests", () => {
 
   it("RemoveParticipant", async () => {
     // mocks
-    const removeParticipantResultMock: RemoveParticipantResult = {};
+    const removeParticipantResultMock: RemoveParticipantResult = {
+      waitForEventProcessor: async () => {
+        return {} as RemoveParticipantEventResult;
+      },
+    };
     callConnection.removeParticipant.returns(
       new Promise((resolve) => {
         resolve(removeParticipantResultMock);
@@ -241,7 +263,7 @@ describe("CallConnection Unit Tests", () => {
 
     // asserts
     promiseResult
-      .then((result: TransferCallResult) => {
+      .then((result: RemoveParticipantResult) => {
         assert.isNotNull(result);
         assert.isTrue(callConnection.removeParticipant.calledWith(target.targetParticipant));
         assert.equal(result, removeParticipantResultMock);
@@ -271,9 +293,34 @@ describe("CallConnection Unit Tests", () => {
       })
       .catch((error) => console.error(error));
   });
+
+  it("CancelAddParticipant", async () => {
+    const invitationId = "invitationId";
+    const cancelAddParticipantResultMock: CancelAddParticipantResult = {
+      invitationId,
+      waitForEventProcessor: async () => {
+        return {} as CancelAddParticipantEventResult;
+      },
+    };
+    callConnection.cancelAddParticipant.returns(
+      new Promise((resolve) => {
+        resolve(cancelAddParticipantResultMock);
+      })
+    );
+
+    callConnection
+      .cancelAddParticipant(invitationId)
+      .then((result: CancelAddParticipantResult) => {
+        assert.isNotNull(result);
+        assert.isTrue(callConnection.cancelAddParticipant.calledWith(invitationId));
+        assert.equal(result, cancelAddParticipantResultMock);
+        return;
+      })
+      .catch((error) => console.error(error));
+  });
 });
 
-describe("CallConnection Live Tests", function () {
+describe.skip("SKIP test until Javascript is updated with TextProxy.CallConnection Live Tests", function () {
   let recorder: Recorder;
   let callerCallAutomationClient: CallAutomationClient;
   let receiverCallAutomationClient: CallAutomationClient;
@@ -484,5 +531,48 @@ describe("CallConnection Live Tests", function () {
       }
     }
     assert.isTrue(isMuted);
+  }).timeout(90000);
+
+  it("Add a participant cancels add participant request", async function () {
+    testName = this.test?.fullTitle()
+      ? this.test?.fullTitle().replace(/ /g, "_")
+      : "cancel_add_participant";
+    await loadPersistedEvents(testName);
+
+    const callInvite: CallInvite = { targetParticipant: testUser2 };
+    const uniqueId = await serviceBusWithNewCall(testUser, testUser2);
+    const callBackUrl: string = dispatcherCallback + `?q=${uniqueId}`;
+    const result = await callerCallAutomationClient.createCall(callInvite, callBackUrl);
+    const incomingCallContext = await waitForIncomingCallContext(uniqueId, 10000);
+    callConnectionId = result.callConnectionProperties.callConnectionId
+      ? result.callConnectionProperties.callConnectionId
+      : "";
+    assert.isDefined(incomingCallContext);
+    if (incomingCallContext) {
+      await receiverCallAutomationClient.answerCall(incomingCallContext, callBackUrl);
+    }
+    const callConnectedEvent = await waitForEvent("CallConnected", callConnectionId, 10000);
+    assert.isDefined(callConnectedEvent);
+    callConnection = result.callConnection;
+    const testUser3: CommunicationUserIdentifier = await createTestUser(recorder);
+    const participantInvite: CallInvite = { targetParticipant: testUser3 };
+
+    const addResult = await callConnection.addParticipant(participantInvite);
+    assert.isDefined(addResult);
+
+    // ensure invitation is sent out
+    await ((ms: number): Promise<void> => new Promise((resolve) => setTimeout(resolve, ms)))(3000);
+
+    // cancel add participant
+    await callConnection.cancelAddParticipant(addResult.invitationId!);
+
+    const addParticipantCancelledEvent = (await waitForEvent(
+      "AddParticipantCancelled",
+      callConnectionId,
+      10000
+    )) as AddParticipantCancelled;
+
+    assert.isDefined(addParticipantCancelledEvent);
+    assert.equal(addResult.invitationId, addParticipantCancelledEvent?.invitationId);
   }).timeout(90000);
 });