+++ sdk/iothub/arm-iothub/src/models/parameters.ts
@@ -53,7 +53,7 @@ export const $host: OperationURLParameter = {
 export const apiVersion: OperationQueryParameter = {
   parameterPath: "apiVersion",
   mapper: {
-    defaultValue: "2023-06-30-preview",
+    defaultValue: "2023-06-30",
     isConstant: true,
     serializedName: "api-version",
     type: {