+++ sdk/network/arm-network/samples-dev/networkWatchersGetSample.ts
@@ -18,7 +18,7 @@ dotenv.config();
  * This sample demonstrates how to Gets the specified network watcher by resource group.
  *
  * @summary Gets the specified network watcher by resource group.
- * x-ms-original-file: specification/network/resource-manager/Microsoft.Network/stable/2023-04-01/examples/NetworkWatcherGet.json
+ * x-ms-original-file: specification/network/resource-manager/Microsoft.Network/stable/2023-05-01/examples/NetworkWatcherGet.json
  */
 async function getNetworkWatcher() {
   const subscriptionId = process.env["NETWORK_SUBSCRIPTION_ID"] || "subid";