+++ sdk/network/arm-network/samples/v32/javascript/availableServiceAliasesListSample.js
@@ -16,7 +16,7 @@ require("dotenv").config();
  * This sample demonstrates how to Gets all available service aliases for this subscription in this region.
  *
  * @summary Gets all available service aliases for this subscription in this region.
- * x-ms-original-file: specification/network/resource-manager/Microsoft.Network/stable/2023-04-01/examples/AvailableServiceAliasesList.json
+ * x-ms-original-file: specification/network/resource-manager/Microsoft.Network/stable/2023-05-01/examples/AvailableServiceAliasesList.json
  */
 async function getAvailableServiceAliases() {
   const subscriptionId = process.env["NETWORK_SUBSCRIPTION_ID"] || "subId";