+++ sdk/network/arm-network/samples/v32/javascript/networkProfilesListAllSample.js
@@ -16,7 +16,7 @@ require("dotenv").config();
  * This sample demonstrates how to Gets all the network profiles in a subscription.
  *
  * @summary Gets all the network profiles in a subscription.
- * x-ms-original-file: specification/network/resource-manager/Microsoft.Network/stable/2023-04-01/examples/NetworkProfileListAll.json
+ * x-ms-original-file: specification/network/resource-manager/Microsoft.Network/stable/2023-05-01/examples/NetworkProfileListAll.json
  */
 async function listAllNetworkProfiles() {
   const subscriptionId = process.env["NETWORK_SUBSCRIPTION_ID"] || "subid";