+++ sdk/recoveryservicessiterecovery/arm-recoveryservices-siterecovery/samples/v5/javascript/replicationProtectionContainersCreateSample.js
@@ -10,17 +10,21 @@
 // Licensed under the MIT License.
 const { SiteRecoveryManagementClient } = require("@azure/arm-recoveryservices-siterecovery");
 const { DefaultAzureCredential } = require("@azure/identity");
+require("dotenv").config();
 
 /**
  * This sample demonstrates how to Operation to create a protection container.
  *
  * @summary Operation to create a protection container.
- * x-ms-original-file: specification/recoveryservicessiterecovery/resource-manager/Microsoft.RecoveryServices/stable/2022-10-01/examples/ReplicationProtectionContainers_Create.json
+ * x-ms-original-file: specification/recoveryservicessiterecovery/resource-manager/Microsoft.RecoveryServices/stable/2023-06-01/examples/ReplicationProtectionContainers_Create.json
  */
 async function createAProtectionContainer() {
-  const subscriptionId = "c183865e-6077-46f2-a3b1-deb0f4f4650a";
+  const subscriptionId =
+    process.env["RECOVERYSERVICESSITERECOVERY_SUBSCRIPTION_ID"] ||
+    "c183865e-6077-46f2-a3b1-deb0f4f4650a";
   const resourceName = "vault1";
-  const resourceGroupName = "resourceGroupPS1";
+  const resourceGroupName =
+    process.env["RECOVERYSERVICESSITERECOVERY_RESOURCE_GROUP"] || "resourceGroupPS1";
   const fabricName = "cloud1";
   const protectionContainerName = "cloud_6d224fc6-f326-5d35-96de-fbf51efb3179";
   const creationInput = {
@@ -40,4 +44,8 @@ async function createAProtectionContainer() {
   console.log(result);
 }
 
-createAProtectionContainer().catch(console.error);
+async function main() {
+  createAProtectionContainer();
+}
+
+main().catch(console.error);