+++ sdk/recoveryservicessiterecovery/arm-recoveryservices-siterecovery/samples/v5/typescript/src/replicationAlertSettingsGetSample.ts
@@ -10,17 +10,24 @@
 // Licensed under the MIT License.
 import { SiteRecoveryManagementClient } from "@azure/arm-recoveryservices-siterecovery";
 import { DefaultAzureCredential } from "@azure/identity";
+import * as dotenv from "dotenv";
+
+dotenv.config();
 
 /**
  * This sample demonstrates how to Gets the details of the specified email notification(alert) configuration.
  *
  * @summary Gets the details of the specified email notification(alert) configuration.
- * x-ms-original-file: specification/recoveryservicessiterecovery/resource-manager/Microsoft.RecoveryServices/stable/2022-10-01/examples/ReplicationAlertSettings_Get.json
+ * x-ms-original-file: specification/recoveryservicessiterecovery/resource-manager/Microsoft.RecoveryServices/stable/2023-06-01/examples/ReplicationAlertSettings_Get.json
  */
 async function getsAnEmailNotificationAlertConfiguration() {
-  const subscriptionId = "c183865e-6077-46f2-a3b1-deb0f4f4650a";
+  const subscriptionId =
+    process.env["RECOVERYSERVICESSITERECOVERY_SUBSCRIPTION_ID"] ||
+    "c183865e-6077-46f2-a3b1-deb0f4f4650a";
   const resourceName = "vault1";
-  const resourceGroupName = "resourceGroupPS1";
+  const resourceGroupName =
+    process.env["RECOVERYSERVICESSITERECOVERY_RESOURCE_GROUP"] ||
+    "resourceGroupPS1";
   const alertSettingName = "defaultAlertSetting";
   const credential = new DefaultAzureCredential();
   const client = new SiteRecoveryManagementClient(credential, subscriptionId);
@@ -32,4 +39,8 @@ async function getsAnEmailNotificationAlertConfiguration() {
   console.log(result);
 }
 
-getsAnEmailNotificationAlertConfiguration().catch(console.error);
+async function main() {
+  getsAnEmailNotificationAlertConfiguration();
+}
+
+main().catch(console.error);