+++ sdk/recoveryservicessiterecovery/arm-recoveryservices-siterecovery/samples/v5/typescript/src/replicationProtectionContainerMappingsUpdateSample.ts
@@ -13,17 +13,24 @@ import {
   SiteRecoveryManagementClient
 } from "@azure/arm-recoveryservices-siterecovery";
 import { DefaultAzureCredential } from "@azure/identity";
+import * as dotenv from "dotenv";
+
+dotenv.config();
 
 /**
  * This sample demonstrates how to The operation to update protection container mapping.
  *
  * @summary The operation to update protection container mapping.
- * x-ms-original-file: specification/recoveryservicessiterecovery/resource-manager/Microsoft.RecoveryServices/stable/2022-10-01/examples/ReplicationProtectionContainerMappings_Update.json
+ * x-ms-original-file: specification/recoveryservicessiterecovery/resource-manager/Microsoft.RecoveryServices/stable/2023-06-01/examples/ReplicationProtectionContainerMappings_Update.json
  */
 async function updateProtectionContainerMapping() {
-  const subscriptionId = "c183865e-6077-46f2-a3b1-deb0f4f4650a";
+  const subscriptionId =
+    process.env["RECOVERYSERVICESSITERECOVERY_SUBSCRIPTION_ID"] ||
+    "c183865e-6077-46f2-a3b1-deb0f4f4650a";
   const resourceName = "vault1";
-  const resourceGroupName = "resourceGroupPS1";
+  const resourceGroupName =
+    process.env["RECOVERYSERVICESSITERECOVERY_RESOURCE_GROUP"] ||
+    "resourceGroupPS1";
   const fabricName = "cloud1";
   const protectionContainerName = "cloud_6d224fc6-f326-5d35-96de-fbf51efb3179";
   const mappingName = "cloud1protectionprofile1";
@@ -50,4 +57,8 @@ async function updateProtectionContainerMapping() {
   console.log(result);
 }
 
-updateProtectionContainerMapping().catch(console.error);
+async function main() {
+  updateProtectionContainerMapping();
+}
+
+main().catch(console.error);