+++ sdk/appconfiguration/app-configuration/CHANGELOG.md
@@ -5,9 +5,16 @@
 
 ### Features Added
 
+- Allow setting `updateIntervalInMs` in `CreateSnapshotOptions`
+
 ### Bugs Fixed
 
 ### Other Changes
+
+- Rename `Snapshot` -> `ConfigurationSnapshot`
+- Rename the properties `retentionPeriod` -> `retentionPeriodInSeconds`, `size` -> `sizeInBytes` for `ConfigurationSnapshot`
+- Rename the type `CompositionType` -> `SnapshotComposition`, `SnapshotStatus` -> `ConfigurationSnapshotStatus`
+- Update the method signature for `archiveSnapshot` and `updateSnapshot` to take in a snapshot name as a parameter
 
 ## 1.5.0-beta.1 (2023-07-14)